# docs/flow-fic-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlowFic API",
    "version": "2026-03-16",
    "description": "OpenAPI spec generated from route handlers in src/app/api and supabase/functions for import into Apidog."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Next.js app"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production Next.js app"
    }
  ],
...
  "paths": {
...
    "/api/writer/earnings/statement": {
      "get": {
        "tags": [
          "Writer Payouts"
        ],
        "summary": "Fetch creator earnings statement",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "storyId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/writer/payout-profile": {
      "get": {
        "tags": [
          "Writer Payouts"
        ],
        "summary": "Get writer payout profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Writer Payouts"
        ],
        "summary": "Create or update writer payout profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalName": {
                    "type": "string"
                  },
                  "promptpayTarget": {
                    "type": "string"
                  }
                },
                "required": [
                  "legalName",
                  "promptpayTarget"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/writer/payouts/request": {
      "post": {
        "tags": [
          "Writer Payouts"
        ],
        "summary": "Submit payout request for current writer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amountSatang": {
                    "type": "integer"
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Business rule error"
          }
        }
      }
    },
    "/api/writer/payouts/history": {
      "get": {
        "tags": [
          "Writer Payouts"
        ],
        "summary": "Fetch payout request history for current writer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SupabaseBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token in the Authorization header."
      },
      "CronBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "CRON_SECRET in the Authorization header for internal cron-protected finance endpoints."
      },
      "StripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature over the raw request body."
      }
    },
...
  }
}